requirements.txt
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
numpy==2.4.2
pillow==12.1.1
pyaes==1.6.1
pyasn1==0.6.2
rsa==4.9.1
setuptools==82.0.0
Telethon==1.42.0
